policy: Add support for port ranges
Add new end_port field to NPDS proto and implement support in
cilium/network_policy.cc.

Check that wildcard port entries do not have an end_port. If our
implementation supported overlapping port ranges we could replace the
special 0 port lookup (for wildcard port) with a 0-65535 range for the
wildcard port.

Change the port parameter in C++ functions to uint16_t as port numbers
are 16 bits. Port in protobuf is still uint32, as there is no uint16 type
in protobufs.

Add the new field to existing tests in various configurations and add a
new tests/cilium_network_policy_test.cc to cover all the edge cases.

+++ cilium/network_policy.h
@@ -55,16 +55,16 @@ class PolicyInstance {
 public:
   virtual ~PolicyInstance() = default;
 
-  virtual bool Allowed(bool ingress, uint32_t port, uint64_t remote_id,
+  virtual bool Allowed(bool ingress, uint16_t port, uint64_t remote_id,
                        Envoy::Http::RequestHeaderMap& headers,
                        Cilium::AccessLog::Entry& log_entry) const PURE;
 
-  virtual const PortPolicyConstSharedPtr findPortPolicy(bool ingress, uint32_t port,
+  virtual const PortPolicyConstSharedPtr findPortPolicy(bool ingress, uint16_t port,
                                                         uint64_t remote_id) const PURE;
 
   // Returns true if the policy specifies l7 protocol for the connection, and
   // returns the l7 protocol string in 'l7_proto'
-  virtual bool useProxylib(bool ingress, uint32_t port, uint64_t remote_id,
+  virtual bool useProxylib(bool ingress, uint16_t port, uint64_t remote_id,
                            std::string& l7_proto) const PURE;
 
   virtual const std::string& conntrackName() const PURE;
